requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiosqlite==0.17.0
async-generator==1.10
async-timeout==3.0.1
asyncio-extras==1.3.2
asyncpraw==7.4.0
asyncprawcore==2.3.0
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
discord.py==1.7.3
emoji==1.4.2
google-auth==1.34.0
google-auth-oauthlib==0.4.5
gspread==4.0.0
idna==3.2
multidict==5.1.0
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.19.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.0
update-checker==0.18.0
urllib3==1.26.6
yarl==1.6.3
